Add files via upload
This project aims to create a dynamic and secure C2C platform using MySQL as the underlying database management system. It involves designing and implementing features that enhance user experience, ensure data security, and enable efficient transactions. The use of MySQL is crucial for managing the vast amounts of data associated with user accounts, product listings, transactions, and reviews in a scalable and reliable manner.